# Conflicts:
#	primefaces/src/main/java/org/primefaces/virusscan/impl/ClamDaemonScanner.java

+++ primefaces/src/main/java/org/primefaces/model/StreamedContent.java
@@ -1,7 +1,7 @@
 /*
  * The MIT License
  *
- * Copyright (c) 2009-2022 PrimeTek
+ * Copyright (c) 2009-2023 PrimeTek Informatics
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -36,7 +36,7 @@ public interface StreamedContent {
 
     String getContentEncoding();
 
-    Integer getContentLength();
+    Long getContentLength();
 
     /**
      * {@link InputStream} which will be buffered and written to the response.
